+++ public_transport_watcher/extractor/insert/transport.py
@@ -1,7 +1,7 @@
 import pandas as pd
 from sqlalchemy.orm import sessionmaker
 
-from public_transport_watcher.db.models import Transport, Categ
+from public_transport_watcher.db.models import Categ, Transport
 from public_transport_watcher.logging_config import get_logger
 from public_transport_watcher.utils import get_engine