+++ CHANGELOG.md
@@ -5,12 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-
 ## [0.9.0] - 2025-04-17
 MR #10
 
 ### Added
-- Extraction and insertion of journey_id
+- Extraction and insertion of `journey_id` for schedules
 
 ## [0.8.0] - 2025-04-17
 MR #9